fix(agents): clear CI blockers for identity update PR
Narrow agent assertions with expectDefined, split UI identity
tests under max-lines, and regenerate Swift protocol models after
nullable emoji/avatar schema.
