+++ ui/src/pages/agents/view.test.ts
@@ -103,65 +103,6 @@ describe("renderAgents", () => {
     expect(onSelectPanel).toHaveBeenCalledWith("tools");
   });
 
-  it("prefills the identity editor from the fetched agent identity", () => {
-    const container = document.createElement("div");
-    render(
-      renderAgents(
-        createProps({
-          agentIdentityById: {
-            beta: { agentId: "beta", name: "Fetched Beta", avatar: "", emoji: "🦊" },
-          },
-        }),
-      ),
-      container,
-    );
-
-    expect(
-      container.querySelector<HTMLInputElement>(".agent-identity-editor__fields input")?.value,
-    ).toBe("Fetched Beta");
-    expect(container.querySelector(".agent-identity-editor__avatar-text")?.textContent).toBe("🦊");
-  });
-
-  it("offers Remove for a stored non-previewable avatar", () => {
-    const onIdentityAvatarClear = vi.fn();
-    const container = document.createElement("div");
-    render(
-      renderAgents(
-        createProps({
-          agentsList: {
-            defaultId: "alpha",
-            mainKey: "main",
-            scope: "workspace",
-            agents: [
-              { id: "alpha", name: "Alpha" } as never,
-              {
-                id: "beta",
-                name: "Beta",
-                identity: { avatar: "https://example.com/avatar.png" },
-              } as never,
-            ],
-          },
-          agentIdentityById: {
-            beta: {
-              agentId: "beta",
-              name: "Beta",
-              avatar: "https://example.com/avatar.png",
-            },
-          },
-          onIdentityAvatarClear,
-        }),
-      ),
-      container,
-    );
-
-    const remove = container.querySelector<HTMLButtonElement>(
-      '[data-testid="agent-identity-avatar-remove"]',
-    );
-    expect(remove).not.toBeNull();
-    remove?.click();
-    expect(onIdentityAvatarClear).toHaveBeenCalledOnce();
-  });
-
   it("shows a model-catalog failure and lets the operator retry", () => {
     const container = document.createElement("div");
     const onModelCatalogRetry = vi.fn();
